pyproject.toml
[tool.poetry]
name = "smnt"
version = "0.1.0"
description = "SMN or SMN't?"
authors = ["Joaquin Torre <joa.dev@live.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.4.3"
requests = "^2.28.1"
ipykernel = "^6.15.1"
SQLAlchemy = "^1.4.39"
prefect = "<2"
apache-airflow = "^2.3.3"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]
pylint = "^2.14.4"
isort = "^5.10.1"
black = "^22.6.0"
pre-commit = "^2.20.0"
flake8 = "^5.0.0"
mypy = "^0.971"
sqlalchemy-stubs = "^0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
line_length = 88
indent = '    '
multi_line_output = 3
length_sort = true
use_parentheses = true
include_trailing_comma = true
force_sort_within_sections = true
known_first_party="smnt"

